refactor WidgetListBox and write test code
convert ListBox to WidgetListBox.
Because I think that doesn't reuse ListBox. So It is changed specific name
