--- app/components/ListBox/ListBox.scss
@@ -1,13 +1,0 @@
-.ListBox {
-  display: flex;
-  flex-direction: column;
-  flex-wrap: nowrap;
-  align-items: stretch;
-
-  border-radius: .25rem;
-  border: 1px solid rgba(0,0,0,.125);
-  font-family: -apple-system,BlinkMacSystemFont,Segoe UI,Roboto,Helvetica Neue,Arial,sans-serif;
-  overflow: scroll;
-  height: 200px;
-  min-width: 350px;
-}
