+++ app/components/WidgetList/WidgetList.scss
@@ -19,13 +19,12 @@
 }
 
 .WidgetList__add-btn {
-  background-color: #e0a800;
-  border: 1px solid rgba(0, 0, 0, .125);
-  width: 100%;
+  background-color: #0000ff;
+  border-radius: 5px;
+  color: #ffffff;
   height: 45px;
   font-size: 1rem;
   font-weight: 400;
   line-height: 1.5;
   padding: .75rem 1.25rem;
-  min-width: 350px;
 }
